添加feign中，crud

+++ src/main/java/cn/com/pojo/User.java
@@ -8,14 +8,15 @@ public class User implements Serializable{
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
-	private Long id;
+	private Integer id;
 	private String UserName;
 	private int level;
 	private String note;
-	public Long getId() {
+	
+	public Integer getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 	public String getUserName() {
